Adaptive Shot Allocation Algorithm (#162)
Adding an implementation of an adaptive shot allocation strategy for quantum expectation value calculations to the braket.experimental.algorithms package. The algorithm is based on the paper "Adaptive estimation of quantum observables" (https://doi.org/10.22331/q-2023-01-26-906) .
Complementing the implementation are two notebooks: an introductory one discussing the importance of shot-allocation; and an example one focusing on the specific algorithm implementation.